+++ restapi/src/main/java/org/sysprotec/restapi/repository/VersionRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface VersionRepository extends JpaRepository<Version, Integer> {
-    Optional<Version> findVersionById(Integer id);
+public interface VersionRepository extends JpaRepository<Version, Long> {
+    Optional<Version> findVersionById(Long id);
     Version findTopByOrderByIdDesc();
-    List<Version> findVersionsByProjectIdOrderByVersionAsc(Integer projectId);
+    List<Version> findVersionsByProjectIdOrderByVersionAsc(Long projectId);
 }
