+++ restapi/src/main/java/org/sysprotec/restapi/repository/VersionStationRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 
 @Repository
-public interface VersionStationRepository extends JpaRepository<VersionStation, Integer> {
-    List<VersionStation> findVersionStationsByVersionIdOrderByIdAsc(Integer versionStationId);
+public interface VersionStationRepository extends JpaRepository<VersionStation, Long> {
+    List<VersionStation> findVersionStationsByVersionIdOrderByIdAsc(Long versionStationId);
 }
