+++ restapi/src/main/java/org/sysprotec/restapi/repository/settings/LopSettingRepository.java
@@ -7,8 +7,8 @@ import org.sysprotec.restapi.model.settings.LopSetting;
 import java.util.Optional;
 
 @Repository
-public interface LopSettingRepository extends JpaRepository<LopSetting, Integer> {
+public interface LopSettingRepository extends JpaRepository<LopSetting, Long> {
 
-    Optional<LopSetting> findLopById(Integer Id);
+    Optional<LopSetting> findLopById(Long Id);
     LopSetting findTopByOrderByIdDesc();
 }
